Add list context return to Win32::HttpGetFile
Don't croak or warn, but make the status message available along with
the return status when called in list context.  In scalar context, we'll
still just get thumbs up or thumbs down, but can get some idea of what
went wrong on failure from Win32::GetLastError(), although we won't know
whether that's a system error or WinHttp error.

Win32::HttpGetFile optionally ignores certificate errors

Make Win32::HttpGetFile fail on HTTP failure

There's not much point in successfully downloading an error page from
the server and saving it to a file, and in fact it's a bit unfriendly
because the user would be unlikely to know there is a problem until
inspecting file contents some time later.