Check bad cert handling in HttpGetFile.t
We'll use a couple of badssl.com sites currently supported by
Google.  The docs say they are for manual testing only, but that's
essentially what we are doing since we don't normally run internet-
connecting tests without manually enabling them.  So we'll only hit
those sites when a developer is doing something extra to test these
features, not when anyone is just installing the Win32 package.

The docs also say these sites are subject to change without notice,
so if that happens, we may have to look elsewhere.